chore: renamed testUtils folder
