+++ MergerLogicUnitTests/DataTypes/FSTest.cs
@@ -1,7 +1,7 @@
 ﻿using MergerLogic.Batching;
 using MergerLogic.DataTypes;
 using MergerLogic.Utils;
-using MergerLogicUnitTests.utils;
+using MergerLogicUnitTests.testUtils;
 using Microsoft.Extensions.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
