+++ MergerLogicUnitTests/DataTypes/S3Test.cs
@@ -3,7 +3,7 @@ using Amazon.S3.Model;
 using MergerLogic.Batching;
 using MergerLogic.DataTypes;
 using MergerLogic.Utils;
-using MergerLogicUnitTests.utils;
+using MergerLogicUnitTests.testUtils;
 using Microsoft.Extensions.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
