+++ MergerLogicUnitTests/testUtils/EqualityComparerFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace MergerLogicUnitTests.utils
+namespace MergerLogicUnitTests.testUtils
 {
     internal static class EqualityComparerFactory
     {
